Move OpenAI adapters into openai-frontend
